Renamed 'Dead By Daylight Mod Installer' project into 'DeadByDaylightModInstaller'. Added licenses
