--- Dead By Daylight Mod Installer/Dead By Daylight Mod Installer.csproj
@@ -1,126 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}</ProjectGuid>
-    <OutputType>WinExe</OutputType>
-    <RootNamespace>Dead_By_Daylight_Mod_Installer</RootNamespace>
-    <AssemblyName>Dead By Daylight Mod Installer</AssemblyName>
-    <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-    <Deterministic>true</Deterministic>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup>
-    <ApplicationIcon>DBM-Logo.ico</ApplicationIcon>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="Newtonsoft.Json, Version=13.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\packages\Newtonsoft.Json.13.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Deployment" />
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Windows.Forms" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Enums\PickResult.cs" />
-    <Compile Include="Model\ISettingsRepository.cs" />
-    <Compile Include="Model\ModPackage.cs" />
-    <Compile Include="Model\Settings.cs" />
-    <Compile Include="Model\SettingsRepository.cs" />
-    <Compile Include="Presenter\CreatorPresenter.cs" />
-    <Compile Include="Presenter\InstallerPresenter.cs" />
-    <Compile Include="Services\Interfaces\IPatcherService.cs" />
-    <Compile Include="Services\PatcherService.cs" />
-    <Compile Include="Services\PickerService.cs" />
-    <Compile Include="Services\Interfaces\IPickerService.cs" />
-    <Compile Include="Services\Interfaces\IMessageBoxService.cs" />
-    <Compile Include="Services\MessageBoxService.cs" />
-    <Compile Include="View\ICreatorView.cs" />
-    <Compile Include="View\IInstallerView.cs" />
-    <Compile Include="View\CreatorForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="View\CreatorForm.Designer.cs">
-      <DependentUpon>CreatorForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="View\InstallerForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="View\InstallerForm.Designer.cs">
-      <DependentUpon>InstallerForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Program.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="View\CreatorForm.resx">
-      <DependentUpon>CreatorForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="View\InstallerForm.resx">
-      <DependentUpon>InstallerForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
-    <Compile Include="Properties\Resources.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <None Include="packages.config" />
-    <None Include="Properties\Settings.settings">
-      <Generator>SettingsSingleFileGenerator</Generator>
-      <LastGenOutput>Settings.Designer.cs</LastGenOutput>
-    </None>
-    <Compile Include="Properties\Settings.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Settings.settings</DependentUpon>
-      <DesignTimeSharedInput>True</DesignTimeSharedInput>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="App.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="background.png" />
-    <None Include="background_creator.png" />
-    <Content Include="DBM-Logo.ico" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\BinaryPatcher\BinaryPatcher\BinaryPatcher.csproj">
-      <Project>{0411c4d1-caec-4ce5-933d-fad1967ec13c}</Project>
-      <Name>BinaryPatcher</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-</Project>
