+++ DeadByDaylightModInstaller.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31613.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeadByDaylightModInstaller", "DeadByDaylightModInstaller\DeadByDaylightModInstaller.csproj", "{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BinaryPatcher", "BinaryPatcher\BinaryPatcher\BinaryPatcher.csproj", "{0411C4D1-CAEC-4CE5-933D-FAD1967EC13C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BCA1B2F3-3DF4-48B0-9BE6-BBE5F146BAD2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0411C4D1-CAEC-4CE5-933D-FAD1967EC13C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0411C4D1-CAEC-4CE5-933D-FAD1967EC13C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0411C4D1-CAEC-4CE5-933D-FAD1967EC13C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0411C4D1-CAEC-4CE5-933D-FAD1967EC13C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A17A3926-08F2-452F-BCBC-B110FE87DB66}
+	EndGlobalSection
+EndGlobal
