+++ DeadByDaylightModInstaller/Enums/PickResult.cs
