+++ DeadByDaylightModInstaller/Model/ISettingsRepository.cs
