+++ DeadByDaylightModInstaller/Model/Settings.cs
