+++ DeadByDaylightModInstaller/Model/SettingsRepository.cs
