+++ DeadByDaylightModInstaller/Presenter/InstallerPresenter.cs
