+++ DeadByDaylightModInstaller/Services/Interfaces/IPatcherService.cs
