+++ DeadByDaylightModInstaller/Services/PatcherService.cs
