+++ DeadByDaylightModInstaller/View/ICreatorView.cs
