+++ DeadByDaylightModInstaller/View/IInstallerView.cs
