+++ DeadByDaylightModInstaller/View/InstallerForm.cs
