+++ DeadByDaylightModInstaller/View/InstallerForm.resx
