+++ DeadByDaylightModInstaller/settings.json
